@@ -1,25 +1,27 @@
 package com.kruczek.theravensystem.rss.source;
 
-import com.kruczek.theravensystem.config.CustomSessionManager;
-import com.kruczek.theravensystem.rss.RssCategory;
-import com.kruczek.theravensystem.rss.entity.RssChannel;
+import java.util.Arrays;
+import java.util.EnumMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import javax.transaction.Transactional;
+
 import org.hibernate.Session;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import javax.transaction.Transactional;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
+import com.kruczek.theravensystem.config.CustomSessionManager;
+import com.kruczek.theravensystem.rss.RssCategory;
+import com.kruczek.theravensystem.rss.entity.RssChannel;
 
 @Repository
 @Transactional
 public class DatabaseRssChannelSource implements RssChannelSource {
 
-    private CustomSessionManager sessionManager;
+    private final CustomSessionManager sessionManager;
 
     @Autowired
     public DatabaseRssChannelSource(CustomSessionManager sessionManager) {
@@ -33,14 +35,22 @@ public class DatabaseRssChannelSource implements RssChannelSource {
 
     @Override
     public void addChannel(String url, RssCategory category) {
-        // TODO: Zrób to zjebie
-        throw new UnsupportedOperationException();
+        RssChannel newRssChannel = new RssChannel();
+        newRssChannel.setUrl(url);
+        newRssChannel.setRssType(category);
+
+        Session session = sessionManager.getSession();
+        session.save(newRssChannel);
     }
 
     @Override
     public void deleteChannel(String url) {
-        // TODO: Zrób to zjebie
-        throw new UnsupportedOperationException();
+        Session session = sessionManager.getSession();
+        RssChannel rssToDelete = (RssChannel) session.createCriteria(RssChannel.class)
+                .add(Restrictions.eq("url", url))
+                .uniqueResult();
+
+        session.delete(rssToDelete);
     }
 
     @Override
@@ -49,7 +59,7 @@ public class DatabaseRssChannelSource implements RssChannelSource {
         Session session = sessionManager.getSession();
         List<RssChannel> rssChannels = session.createCriteria(RssChannel.class).list();
 
-        Map<RssCategory, List<String>> result = new HashMap<>();
+        Map<RssCategory, List<String>> result = new EnumMap<>(RssCategory.class);
 
         for (RssChannel rssChannel : rssChannels) {
             RssCategory rssType = rssChannel.getRssType();
