@@ -1,9 +1,9 @@
 package com.kruczek.theravensystem.rss.source;
 
-import com.kruczek.theravensystem.rss.RssCategory;
-
 import java.util.List;
 import java.util.Map;
+
+import com.kruczek.theravensystem.rss.RssCategory;
 
 public interface RssChannelSource {
 
